(Import Cleanup) - Moved `time` into `ts.py` - Cleaned up imports
