--- hcm/signal/__init__.py
@@ -1,2 +1,0 @@
-from hcm.signal.time import time
-from hcm.signal.vco import VCO
