+++ hcm/signal/vc.py
@@ -1,5 +1,5 @@
 import numpy as np
-import hcm.ts.time as time
+import hcm.ts
 
 # allow voltage control of osc frequency
 # here freq is now an array w/ same length as t
@@ -26,9 +26,9 @@ def VCA(signal, modulation):
 
 def ADSR(A, D, S, R, duration, sample_rate):
     """TODO: Find good place for this... """
-    attack = (1.0 - 0.0) / A * time(0, A, sample_rate)
-    decay = 1 - S / D * time(0, D, sample_rate)
-    release = S - (S - 0.0) / R * time(0, R, sample_rate)
+    attack = (1.0 - 0.0) / A * hcm.ts.time(0, A, sample_rate)
+    decay = 1 - S / D * hcm.ts.time(0, D, sample_rate)
+    release = S - (S - 0.0) / R * hcm.ts.time(0, R, sample_rate)
     sustain = S * np.ones(int(duration * sample_rate) - len(attack) - len(decay) - len(release))
 
     return np.concatenate([attack, decay, sustain, release])
